requirements.txt
absl-py==0.14.0
beautifulsoup4==4.10.0
bs4==0.0.1
finance-datareader==0.9.31
matplotlib==3.4.3
pykrx
